mrc-ngram-embedding-manager(nem-migrate-maybe-prepare-move,non_mechanical_provable): Move prepare_for_forward onto NgramEmbeddingManager (cut+paste)
