--- python/sglang/srt/layers/n_gram_embedding_manager.py
@@ -1,85 +1,0 @@
-from __future__ import annotations
-
-from dataclasses import dataclass
-from typing import TYPE_CHECKING, Optional
-
-import torch
-
-from sglang.srt.configs.model_config import ModelConfig
-from sglang.srt.mem_cache.memory_pool import ReqToTokenPool
-from sglang.srt.model_executor.ngram_token_table import (
-    update_ngram_token_table_after_sampling,
-)
-from sglang.srt.server_args import ServerArgs
-
-if TYPE_CHECKING:
-    from sglang.srt.model_executor.forward_batch_info import ForwardBatch
-
-
-@dataclass(frozen=True, slots=True, kw_only=True)
-class NgramEmbeddingManager:
-    enabled: bool
-    table: Optional[torch.Tensor]
-    n: int
-    k: int
-
-    @classmethod
-    def from_model(
-        cls,
-        *,
-        model: torch.nn.Module,
-        model_config: ModelConfig,
-        req_to_token_pool: ReqToTokenPool,
-        server_args: ServerArgs,
-        max_running_requests: int,
-        device: str,
-    ):
-        token_table = None
-        ngram_embedding_n = 0
-        ngram_embedding_k = 0
-        use_ngram_embedding = model_config.use_ngram_embedding
-        if use_ngram_embedding:
-            from sglang.srt.layers.n_gram_embedding import NgramEmbedding
-
-            # Sized to mirror req_to_token (indexed by req_pool_idx).
-            token_table = torch.empty(
-                req_to_token_pool.req_to_token.shape[0],
-                model_config.context_len,
-                dtype=torch.int32,
-                device=device,
-            )
-            chunked_prefill_size = server_args.chunked_prefill_size
-            assert (
-                chunked_prefill_size is not None and chunked_prefill_size > 0
-            ), "Ngram embedding requires chunked prefill to be enabled (chunked_prefill_size > 0)"
-            for module in model.modules():
-                if isinstance(module, NgramEmbedding):
-                    module.init_buffers(
-                        max_running_requests, chunked_prefill_size, device
-                    )
-            hf_config = model_config.hf_config
-            ngram_embedding_n = hf_config.ngram_embedding_n
-            ngram_embedding_k = hf_config.ngram_embedding_k
-        return cls(
-            enabled=use_ngram_embedding,
-            table=token_table,
-            n=ngram_embedding_n,
-            k=ngram_embedding_k,
-        )
-
-    def update_after_decode(
-        self,
-        next_token_ids: torch.Tensor,
-        forward_batch: ForwardBatch,
-    ):
-        """Update the ngram embedding token table after sampling."""
-        ngram_embedding_info = forward_batch.ngram_embedding_info
-        if ngram_embedding_info is None:
-            return
-        update_ngram_token_table_after_sampling(
-            ngram_embedding_info=ngram_embedding_info,
-            next_token_ids=next_token_ids,
-            req_pool_indices=forward_batch.req_pool_indices,
-            seq_lens=forward_batch.seq_lens,
-            batch_size=forward_batch.batch_size,
-        )
